+++ app/src/main/java/com/example/etherealartefacts/EtherealArtefactsApp.kt
@@ -4,5 +4,5 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class App: Application() {
+class EtherealArtefactsApp: Application() {
 }
