+++ app/src/main/java/com/example/etherealartefacts/main/MainActivity.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterial3Api::class)
 
-package com.example.etherealartefacts
+package com.example.etherealartefacts.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -16,6 +16,8 @@ import com.example.etherealartefacts.ui.theme.EtherealArtefactsTheme
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.etherealartefacts.ui.login.LoginScreen
+import com.example.etherealartefacts.ui.productsDetails.ProductsScreen
 import com.example.etherealartefacts.networking.JWTTokenProvider
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
