+++ app/src/main/java/com/example/etherealartefacts/ui/login/LoginScreen.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalComposeUiApi::class, ExperimentalMaterial3Api::class)
 
-package com.example.etherealartefacts
+package com.example.etherealartefacts.ui.login
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
@@ -41,6 +41,7 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.example.etherealartefacts.R
 import com.example.etherealartefacts.models.LoginRequest
 import com.example.etherealartefacts.models.LoginResponse
 import com.example.etherealartefacts.networking.JWTTokenProvider
