+++ app/src/main/java/com/example/etherealartefacts/ui/login/LoginViewModel.kt
@@ -1,8 +1,10 @@
-package com.example.etherealartefacts
+package com.example.etherealartefacts.ui.login
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.etherealartefacts.models.ProductDetailsModel
+import com.example.etherealartefacts.models.LoginRequest
+import com.example.etherealartefacts.models.LoginResponse
+import com.example.etherealartefacts.repository.DefaultRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -10,25 +12,23 @@ import kotlinx.coroutines.launch
 import java.lang.Exception
 import javax.inject.Inject
 
-
 @HiltViewModel
-class ProductDetailsViewModel @Inject constructor(
+class LoginViewModel @Inject constructor(
     private val repository: DefaultRepository,
+) : ViewModel() {
 
-    ) : ViewModel() {
-
-    private val _response = MutableStateFlow<Result<ProductDetailsModel>?>(null)
-    val response : StateFlow<Result<ProductDetailsModel>?> = _response
+    private val _response = MutableStateFlow<Result<LoginResponse>?>(null)
+    val response: StateFlow<Result<LoginResponse>?> = _response
 
     private val _isLoading = MutableStateFlow(false)
     val isLoading: StateFlow<Boolean> = _isLoading
 
-    fun getProductDetails(id: Int) {
+    fun login(request: LoginRequest) {
         viewModelScope.launch {
             try {
                 _isLoading.value = true
-                val getProductDetailsResponse = repository.getProductDetails(id)
-                _response.value = Result.success(getProductDetailsResponse)
+                val loginResponse = repository.login(request)
+                _response.value = Result.success(loginResponse)
             } catch (e: Exception) {
                 _response.value = Result.failure(e)
             } finally {
