+++ app/src/main/java/com/example/etherealartefacts/ui/productsDetails/ProductDetailsViewModel.kt
@@ -1,9 +1,9 @@
-package com.example.etherealartefacts
+package com.example.etherealartefacts.ui.productsDetails
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.etherealartefacts.models.LoginRequest
-import com.example.etherealartefacts.models.LoginResponse
+import com.example.etherealartefacts.models.ProductDetailsModel
+import com.example.etherealartefacts.repository.DefaultRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -11,23 +11,25 @@ import kotlinx.coroutines.launch
 import java.lang.Exception
 import javax.inject.Inject
 
-@HiltViewModel
-class LoginViewModel @Inject constructor(
-    private val repository: DefaultRepository,
-) : ViewModel() {
 
-    private val _response = MutableStateFlow<Result<LoginResponse>?>(null)
-    val response: StateFlow<Result<LoginResponse>?> = _response
+@HiltViewModel
+class ProductDetailsViewModel @Inject constructor(
+    private val repository: DefaultRepository,
+
+    ) : ViewModel() {
+
+    private val _response = MutableStateFlow<Result<ProductDetailsModel>?>(null)
+    val response : StateFlow<Result<ProductDetailsModel>?> = _response
 
     private val _isLoading = MutableStateFlow(false)
     val isLoading: StateFlow<Boolean> = _isLoading
 
-    fun login(request: LoginRequest) {
+    fun getProductDetails(id: Int) {
         viewModelScope.launch {
             try {
                 _isLoading.value = true
-                val loginResponse = repository.login(request)
-                _response.value = Result.success(loginResponse)
+                val getProductDetailsResponse = repository.getProductDetails(id)
+                _response.value = Result.success(getProductDetailsResponse)
             } catch (e: Exception) {
                 _response.value = Result.failure(e)
             } finally {
